Refactor enemy entity retrieval in CollisionSystem for improved context handling
- Removed the previous synchronous `getEnemyEntities` method and replaced it with a context-aware version that supports cancellation.
- Updated the `checkPlayerEnemyCollisions` method to handle errors from the new `getEnemyEntities` method, enhancing error management during collision checks.
- Streamlined the retrieval of enemy entities to include health information, improving the accuracy of collision detection.
