+++ internal/ecs/systems/collision/system.go
@@ -5,9 +5,12 @@ import (
 	"time"
 
 	"github.com/yohamta/donburi"
+	"github.com/yohamta/donburi/filter"
+	"github.com/yohamta/donburi/query"
 
 	"github.com/jonesrussell/gimbal/internal/common"
 	"github.com/jonesrussell/gimbal/internal/config"
+	"github.com/jonesrussell/gimbal/internal/ecs/core"
 	"github.com/jonesrussell/gimbal/internal/ecs/managers"
 )
 
@@ -77,4 +80,27 @@ func (cs *CollisionSystem) GetCollisionDistance(pos1, pos2 common.Point) float64
 func (cs *CollisionSystem) IsWithinRange(pos1, pos2 common.Point, maxDistance float64) bool {
 	distance := cs.GetCollisionDistance(pos1, pos2)
 	return distance <= maxDistance*maxDistance // Compare squared distances
+}
+
+// getEnemyEntities returns all valid enemy entities with health
+func (cs *CollisionSystem) getEnemyEntities(ctx context.Context) ([]donburi.Entity, error) {
+	// Check for cancellation
+	select {
+	case <-ctx.Done():
+		return nil, ctx.Err()
+	default:
+	}
+
+	enemies := make([]donburi.Entity, 0)
+	query.NewQuery(
+		filter.And(
+			filter.Contains(core.EnemyTag),
+			filter.Contains(core.Position),
+			filter.Contains(core.Size),
+			filter.Contains(core.Health),
+		),
+	).Each(cs.world, func(entry *donburi.Entry) {
+		enemies = append(enemies, entry.Entity())
+	})
+	return enemies, nil
 }
